작업중

+++ src/main/java/com/example/demo/api/v1/notice/entity/Notice.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @Builder
 public class Notice {
-    private int noticeId;
+    private long noticeId;
     private String noticeTitle;
     private String noticeContent;
     private String deleteYn;
